wear: chore extract simple UI in util class
